Java/src/main/java/com/gildedrose/GildedRose.java
package com.gildedrose;


class GildedRose {

    public static final String BACKSTAGE_PASSES = "Backstage passes to a TAFKAL80ETC concert";
    public static final String SULFARAS = "Sulfuras, Hand of Ragnaros";
    public static final String AGED_BRIE = "Aged Brie";

    Item[] items;

    public GildedRose(Item[] items) {
        this.items = items;
    }

    public void updateQuality() {
        for (Item item : items) {
            final ItemWrapper wrappedItem = ItemWrapper.of(item);
            wrappedItem.update();
        }
    }
}
